Fin Day 14
Today is releaseday again :top:
Changes don't include the actuall playability of the game, simply
because it's very unstable at this point. Furthermore is the theoretical
part of this project underway and should be 100% done until 10.11. the
deadline of my work...
Here's the full changelog included in this update:
*The 'java.library.path' now gets set at runtime -> you won't need to
start the jar with the -Djava..... arguments
*outsourced getID in each state to the base state
*collapsed all method bodies in base state
*added effects and items, still more to come!
*added Inventory management -> leaves posability to make game
multiscreen
*added 'OSManagement' wich tells everything about the users operating
system
*Fixed Pirate Language 'confirm popup' not breaking lines....